v 1.4.6
Add KDU_EPROCESS_OFFSETS and KDUGetEprocessOffsets to centralize EPROCESS offsets (PsProtection and mitigation flags) and refactor ps.* logic to use the new API. Add provider callback verification before performing PS patches. Fix multiple bugs: correct instruction validation logic in dsefix, use explicit size cast in zemana, handle NULL SD case and avoid double-free in ipcsvc, add dbghelp existence check before LoadLibrary in sym, rename TEXT_SECTION_LEGNTH->TEXT_SECTION_LENGTH and use it in victim, and add retry limit for IPC connect. Update many embedded driver/resource binaries accordingly.

+++ Source/Hamakaze/idrv/zemana.cpp
@@ -1,12 +1,12 @@
 /*******************************************************************************
 *
-*  (C) COPYRIGHT AUTHORS, 2022 - 2023
+*  (C) COPYRIGHT AUTHORS, 2022 - 2026
 *
 *  TITLE:       ZEMANA.CPP
 *
-*  VERSION:     1.40
+*  VERSION:     1.46
 *
-*  DATE:        20 Oct 2023
+*  DATE:        12 Feb 2026
 *
 *  Zemana driver routines.
 *
@@ -547,7 +547,7 @@ BOOL WINAPI ZmOpenProcess(
         &ProcessId,
         sizeof(ProcessId),
         ProcessHandle,
-        sizeof(ProcessHandle));
+        (ULONG)sizeof(ProcessHandle));
 }
 
 /*
